requirements.txt
# Core dependencies
python-pptx>=0.6.23
pillow>=10.4.0
PyPDF2>=3.0.0

# Method 1: Playwright (recommended - best rendering)
playwright>=1.40.0

# Method 2: WeasyPrint (pure Python, no browser needed)
weasyprint>=60.0
pdf2image>=1.16.0

# Optional: Progress bar and watch mode
tqdm>=4.66.0
watchdog>=3.0.0

# Optional: Format conversion (PDF ↔ PPT)
reportlab>=4.0.0

# Note: After installing playwright, run:
# playwright install chromium
